increasing_heap_allocator

Trait PageAllocatorProvider

pub trait PageAllocatorProvider<const PAGE_SIZE: usize> {
    // Required methods
    fn allocate_pages(&mut self, pages: usize) -> Option<*mut u8>;
    fn deallocate_pages(&mut self, pages: usize) -> bool;
}

Required Methods

fn allocate_pages(&mut self, pages: usize) -> Option<*mut u8>

Return the start address of the new allocated heap

fn deallocate_pages(&mut self, pages: usize) -> bool

Deallocate pages from the end of the heap Return true if the deallocation was successful